raop: Move RTSP implementation to support
Relates to #1255
